refactor: migration to sentence casing (#4089)
For consistency reasons we decided to move all texts to sentence casing.
This applies to all titles, descriptions, labels and status badges,
basically more or less all texts in our locales files and the texts we
haven't moved to i18n yet. Anything that is served from the backend
remains as is.

Part of https://github.com/kumahq/kuma-gui/issues/3774

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>