feat: export env and can from new settings package (#4113)
Part of the overall restructuring and is a pre-effort to outsource the
testing service into another package.

---

The `CliEnv`, which is only used during unit testing, should not be part
of the common export of the new package and therefore not added to the
barrel file. Otherwise it would be part of the bundle and end up in the
browser. Therefore I've just added a separate export to the manifest.
That's a common problem with barrel files and we should be careful what
is added to the barrel files while we restructure the codebase.

---

Regarding unit tests: since we have `test.setupFiles` added to our base
config I had to create such files in the packages where I started to add
unit tests. I created some placeholder files, but I wasn't quite happy
about it. Instead I've just decided to remove the `test.setupFiles` from
the base config leaving it up to the consumer to set this.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>