fix(deps/security-actions/sca): use hash of version release (#4132)
Our release workflow to the kuma repo is currently failing. It looks
like there is an issue upstream. It started failing since we've used a
hash that is newer than the hash of the actual release of the latest
version of `Kong/public-shared-actions/security-actions/sca`. This is an
attempt to make it work by using the hash of the actual release.

We also missed to add the renovate preset
`Kong/public-shared-renovate:github-actions` to ensure that renovate
will always use the digest of semver version.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>
Signed-off-by: Bart Smykla <bartek@smykla.com>
Co-authored-by: Bart Smykla <bartek@smykla.com>