fix(rest-client): limited support of request.body (#4135)
To send data as payload in our requests (i.e. `POST`, `PATCH`, `PUT`) we
currently use
[Request](https://developer.mozilla.org/en-US/docs/Web/API/Request) and
[Response](https://developer.mozilla.org/en-US/docs/Web/API/Response/Response)
to extract the payload and pass it further to `fetch`. This is part of
the integration of the openapi-fetch and the option to pass an own
`fetch` implementation when using
[`createClient`](https://openapi-ts.dev/openapi-fetch/api#createclient).
To extract the payload from the `Request` we currently use
[`Request.body`](https://developer.mozilla.org/en-US/docs/Web/API/Request/body),
which is not supported when using firefox. Instead we can use
[`Request.blob`](https://developer.mozilla.org/en-US/docs/Web/API/Request/blob),
which is what I did here now.

Note: I tested to replace `req.body` with `req.blob` entirely, but the
logs showed that the blob was empty when not using firefox. It still
seemed to work as the final request in the browser inspector revealed
the correct payload to be part of the request. But I felt it's better to
keep using `req.body` and only use `req.blob` if `req.body` isn't set.
At least this makes it easier to debug.

Signed-off-by: schogges <moritz.fleck@konghq.com>