feat(data-planes): add new data-planes based on dp feature flag (#4133)
Most of the changes are caused by the duplication of `data-planes` to
`legacy-data-planes` to be able to integrate the new APIs without
affecting the previous version. Once the integration is done, we can
entirely remove the `legacy-data-planes` and everything related to it
(i.e. e2e tests and services) and expect very few changes in other
places. This approach is causing a lot of changes on the current
data-planes now, but less on the new data-planes once we are done.

The new data-planes start with less features, such as the
`Inbounds/Outbounds` (traffic) listing. For now he policy tab will be
removed completely (see #4118), we may bring it back at some point but
only for proxy scoped polices.

In order to switch between the two routes (and subroutes) without any
interference I've added a `RootView` and `DataPlaneRouteGuard` to the
`legacy-data-planes`. This loads the `DataplaneOverview` source and
checks for the feature flag being set. Depending on the feature flag the
current routes are being completely replaced and the router loads the
respective routes.

I've added the feature flag to the mocks. In order to enforce the
feature flag being set or not we can set a cookie:
`KUMA_DATAPLANE_RUNTIME_UNIFIED_RESOURCE_NAMING_ENABLED` (`true` or
`false`). If not set it's a random chance to have the feature flag
included or not.

Closes #4118

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>