feat: mute validation errors (#4140)
In some cases we want to validate user inputs. The validation happens in
the data source layer, so we could potentially move the validation to
the backend with few changes necessary. But this also causes errors
being thrown unnecessarily. This PR mutes these errors but still let the
error being displayed to the user, i.e. below an input field that is
being validated.

Signed-off-by: schogges <moritz.fleck@konghq.com>