feat(legacy-data-planes): keep legacy for gateways (#4149)
We realized that we cannot use the `{kri}/_layout` and `{kri}/_policies`
endpoints for gateway proxies. For now gateway dataplane detail views
should basically just keep working as is.
We can't even fully rely on the feature flag being set, we also have to
check for the proxy type.
This makes sure that we show the new views only for standard sidecar
proxies.

Closes https://github.com/kumahq/kuma-gui/issues/4118

Signed-off-by: schogges <moritz.fleck@konghq.com>