feat(data-planes): add traffic summary view (#4141)
This PR creates a new shared root view for dataplane inbounds and
outbounds which renders the tabs in the summary tray. There is not much
of a difference so this can be a shared view. Although both inbound
overview and outbound overview are very similar, there are some minor
difference that makes it worth keeping them as separate views.

I had to make sure that no other tabs are being rendered, as they will
be addressed separately. Therefore I've commented out some bits in the
`routes` file and only add the respective overviews as children.

In order to parse and further use parts of a KRI I've added classes for
full and contextual KRIs (also see [examples in
madr](https://github.com/kumahq/kuma/blob/master/docs/madr/decisions/077-migrating-to-consistent-and-well-defined-naming-for-non-system-envoy-resources-and-stats.md#examples)).

---

Note: Make sure to hide whitespaces when reviewing

Closes https://github.com/kumahq/kuma-gui/issues/4120

Signed-off-by: schogges <moritz.fleck@konghq.com>