feat(data-planes): relocate subscriptions listing (#4150)
The subscriptions listing (XDS connections) was a bit lost down below
the traffic listing (inbounds/outbounds), so we've decided to relocate
it. The most important information is now part of the DP about section
and adds a link to a summary tray that shows the whole listing with all
information.
There is only one active connection to a control plane instance at a
time, so either we show the `connectedTime`, `controlPlaneInstanceId`
and `kumaDp.version` in case the connection is alive or a text that
indicates that there is no active connection. In case there are no
subscriptions at all (neither active nor inactive) the section is
hidden.

Note: As the subscriptions are not related to the new DP feature flag
(`feature-unified-resource-naming`) I've done the same change for both
`data-planes` and `legacy-data-planes`.

Closes https://github.com/kumahq/kuma-gui/issues/4137

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>