feat(mesh): show mesh-identities (#4162)
In order to show the user `MeshIdentity` resources I've added a
collection (as badges) to the mesh about section. If there are no
`MeshIdentity` resource the section is hidden.
Each item of the collection is clickable and opens a summary tray which
shows the config of the `MeshIdentity`.

---

Since `MeshIdentity` is a more or less decoupled resource, I've added it
to the `resources` folder so it could be used in any sub-route of a
mesh.

Closes https://github.com/kumahq/kuma-gui/issues/4160

Signed-off-by: schogges <moritz.fleck@konghq.com>