fix(data-planes/inbound-summary): e2e flake, non-matching inbounds (#4168)
Since adding more info from the dataplane `_overview` endpoint to the
inbound summary view (https://github.com/kumahq/kuma-gui/pull/4157)
there is a flake because we match the inbounds from both endpoints
(`_overview` and `_layout`) by checking the `portName ===
<contextualKri>.sectionName` and `port`. We then show the protocol which
is in `dataplane.networking.inbound[].tags['kuma.io/protocol']`.
This PR makes sure that the inbounds always match and the protocol we
assert for is in the fixture.

Signed-off-by: schogges <moritz.fleck@konghq.com>