fix(data-planes-traffic-summary): wrong kri for fetching policies (#4167)
We used the wrong `kri` when fetching policies of an inbound. We should
not use the contextual kri (i.e. `self_inbound_dp_http`) but the full
kri.
I've also changed it in outbounds to make things work the same, even
though it worked because `kri` and `proxyResourceName` are the same on
outbounds.
I've also changed the title of the the policies section from `Rules` to
`Policies`.

Closes https://github.com/kumahq/kuma-gui/issues/4170

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>