feat(mesh): show mesh-trusts (#4165)
This adds a listing of `MeshTrust` resources to the mesh detail view.
Similar to `MeshIdentity` this is a more or less decoupled resource and
therefore I've added it to the `resources` folder, so it can be used in
any sub-route of a mesh.
The listing is displayed below the about section and clicking on an
entry opens a summary tray which shows the config of the `MeshTrust`
resource.
Each `MeshTrust` has a `MeshIdentity` origin which is linked in the
listing in the third column. Clicking on the kri opens the summary tray
of the `MeshIdentity`.

<img width="1402" height="592" alt="Screenshot 2025-08-20 at 10 28 40"
src="https://github.com/user-attachments/assets/0c4d059e-255f-40ab-bb98-7e86d33cf149"
/>
<img width="1424" height="1016" alt="Screenshot 2025-08-20 at 10 28 48"
src="https://github.com/user-attachments/assets/742bb68c-2c9b-4cbb-99ee-fc2b404938cf"
/>

Closes https://github.com/kumahq/kuma-gui/issues/4161

Signed-off-by: schogges <moritz.fleck@konghq.com>