feat(data-planes): check for mesh-services mode to use unified-naming (#4176)
We've noticed that when `meshServices.mode` is disabled on the mesh, we
can't use `feature-unified-resource-naming`. Until this condition is
fully considered on the backend, we are going to check for the
`meshServices.mode` being set to `Exclusive`. The backend will work on
exposing a new field `dataplaneInsights.features` which will either
include the feature flag or not by taking into account if mesh services
are enabled or not.

Signed-off-by: schogges <moritz.fleck@konghq.com>