feat(data-planes): reflect spire in dp overview (#4174)
In case of spiffe/spire the TLS certificate of a dataplane is issued via
`MeshIdentity` resource and managed externally. In this case we have
less information, such as `lastCertificateRegeneration`,
`certificateExpirationTime` and `supportedBackends` are not set. The
`issuedBackend` is a kri of the `MeshIdentity`.
We then tell the user that "The certificate is managed externally" and
only show the `issuedBackend`. The kri is clickable and opens a summary
tray of the `MeshIdentity` resource.

---

The routes of `MeshIdentity` are now also added to the sub-routes of a
dataplane, such that the user can inspect the `MeshIdentitySummaryView`
without leaving the context of the dataplane. To achieve that, I've
updated the `MeshIdentitySummaryView` to load the data instead of
passing it as props. Also the `routeName` is now dynamic.

Closes https://github.com/kumahq/kuma-gui/issues/4159

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>