fix(e2e/tls): flake timing issue (#4186)
Again one of those scenarios where the click on the link happens too
fast leading to the routing not being performed and the test fails. I've
added a random `wait for 500 ms` to get around this issue.
Furthermore I've deduped the env variable
`KUMA_DATAPLANE_TLS_ISSUED_MESHIDENTITY` as this is already set for all
scenarios.

Signed-off-by: schogges <moritz.fleck@konghq.com>