refactor(data-planes): cleanup legacy-data-planes (#4177)
After implementing all the expected changes to support
`feature-unified-resource-naming` in data-planes, we are now able to
remove all the duplicated code.
Essentially this removes all duplication from `legacy-data-planes` and
only leaves what is different. As of next we are going further to remove
duplication from the individual view files such that we can finally
remove the duplicated routes and the route switching.
We still a more or less duplicated set of e2e tests that test the
data-planes with the feature flag disabled. This will probably stay a
little longer.

Signed-off-by: schogges <moritz.fleck@konghq.com>