feat(docs): several docs links amends (#4189)
The following links are currently out-of-date.

For the `gateway` link I'm "almost sure" that its not even used, but
felt it was easier to replace it anyway. But if anyone can confirm for
me that they can't see it being used in our code also, I'd be happy to
just totally remove it from the locales files. I'm pretty sure its just
left over from an older version. lmk

Signed-off-by: John Cowen <john.cowen@konghq.com>