feat(data-planes): improve mocking regarding unified naming (#4190)
This slightly improves the mocking of `stats` and `_layout` when it
comes to unified names. The key is to make sure generated values like
ports are synced across the two mock handlers. There is still a lot of
randomness.
I've compared the output of a real deployment and turned out there were
some misconfigurations here, such as `port` as `sectionName` in `stats`
and excluding the `namespace` from the `name` when generating the `kri`
for outbounds.
For `inbounds` it was more or less straightforward to add the respective
`listener.`s.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>