fix(types): fixup DataplaneOverview type in DataplaneRouteGuard (#4195)
Whilst working on https://github.com/kumahq/kuma-gui/pull/4194 I found
that one side of `can` was expecting the DataplaneOverview from
`@/app/data-planes/data` and the other side was using our old
DataPlaneOverview from `@/types`.

This makes them the same.

https://github.com/kumahq/kuma-gui/pull/4194 goes some way to prevent
this from happening, although this wasn't the reason for me working on
that.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>
Signed-off-by: John Cowen <johncowen@users.noreply.github.com>
Co-authored-by: Mo <24444046+schogges@users.noreply.github.com>