feat(data-plane): remove link to MeshIdentity (#4204)
In case the issuer for mTLS is `MeshIdentity` we show the `kri` of the
`MeshIdentity` in the DP's TLS section. We use to link from the `kri` to
the `MeshIdentitySummaryView`, which turns out is not possible as we
cannot reconstruct the correct `name` parameter for
`/meshes/{mesh}/meshidentities/{name}` to fetch the config.
So until we can potentially use the full `kri` as the `name` parameter
we have to remove the linking and therefore just display the `kri` as
is.

Signed-off-by: schogges <moritz.fleck@konghq.com>