refactor: remove x-growth-*, improve x-empty-state (#4210)
The `XEmptyState` component is auto-filling some content in case the
correct locales keys are present. It should only support
`x-empty-state`-keys and therefore I've moved everything from
`x-growth-empty-state` to `x-empty-state`.
Also there was a tiny typo of the `types`: it should be `zones-crud`,
not `zone-crud`.
As next steps we also want to remove everything around crud.

Signed-off-by: schogges <moritz.fleck@konghq.com>