refactor(resources/routes): use kri for routing of mid and mtrust (#4198)
We decided to use `kri`s as is whenever possible. The first and probably
easiest places where we can do this is the new resources `MeshIdentity`
and `MeshTrust` which have been lately being added within the context of
`kri`s.
There are still some places that require the `name` instead of the
`kri`, but at the very least we are trying to remove the parsing of
`kri`s from the views and only keep it in the data-layer and sources.

The following routes now expect a `kri`:
- `mesh-identity-summary-view`: `/meshes/:mesh/meshidentities/:mid` -
`:mid` is expected to be a `kri`
- `mesh-trust-summary-view`: `/meshes/:mesh/meshtrusts/:mtrust` -
`:mtrust` is expected to be a `kri`

For both `MeshTrust` and `MeshIdentity` I've added to build a `kri` from
the provided data and add it to the returned data in the data-layer. For
now this is just a minimal `kri`, i.e. `kri_mid_default___identity-1_`
with only `shortName`, `mesh` and the `name` of the resource.

---

Furthermore I've fixed some minor things
- the route name of `MeshTrustSummaryView` was set to
`mesh-identity-summary-view`
- now that we add more data to the response of `MeshTrust` and
`MeshIdentity`, I've added another entry `raw` to preserve the returned
data as config.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>