chore(application): refactor env to use a composable approach (#4211)
Converts `env` to work the same as `can` (and more or less `uri`).

New environment variables are defined as an object of string to
functions, and `declare`d.

These are composed in the service container using a new `$.vars` token.

"Env vars" remain backed by cookies during development and during
testing.

---

Once this is in and we are happy with it, there will probably be a
little more moving around, for example moving `semver` back into the
application instead of `@kumahq/config` plus probably the odd other
thing. Just this PR already needed to be rather large due to the impact
of `env` through our application, so I didn't want to do too much in one
go.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>