refactor(kri): simplify parsing kri (#4188)
I've realized that `proxyResourceNames` are no real `KRI`s. Therefore
I've removed the `ContextualKri` and the inline parsing entirely leaving
just a few places where we have to parse KRIs to extract some
information for data-fetching or routing.
There is a tiny parsing we still need to do about the
`proxyResourceName`, which I've moved to the data-layer. In order to
properly map envoy stats with the `proxyResourceName` of inbounds, we
need to replace the section name with the `port` of an inbound. For the
moment I've named this new entry `proxyResourcePortName`, but we hope to
get this updated in the backend, such that we can also remove this data
modeling.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>