chore(ci): add post release / kuma sync workflow (#4262)
I plan on making a workflow dispatch only workflow for us to use
immediately after a kuma release.

It should at least:

- [x] sync the active-branches.json from the upstream repo
- [ ] depending on what we are doing regarding backports etc, make sure
things are in sync (if we don't need to sync back we don't need to
ensure we do this one)
- [x] ideally automatically do a renovate package-lock maintenance - but
we are already semi-automatic with renovate anyway, so maybe just some
sort of message in the PR we create
- [x] create a PR with the synced active-branches.json

Work towards https://github.com/kumahq/kuma-gui/issues/3710

---

edit:

I kinda struggled with the naming of this seeing as `kuma-gui` does a
release of its code on every PR merge, whereas `kuma` does a release
every few months. They are very different things and I didn't want to
confuse things.

In the end I called this "Sync kuma-gui from kuma" which is super clear
as to what this currently does.

I plan on merging this with only the above two check boxes checked
because:

- its not clear to me what we want to do with backports. Judging by the
amount of changes we need to do and then sync to another branch during
stabilisation, personally I would rather cherrypick the odd change we
make instead of doing a total sync after possibly several weeks. Waiting
for a total sync to happen weeks later can sometimes get in the way of
work on `master` so I'd prefer not to do that. Happy to set up some sort
of
[automation](https://github.com/kumahq/kuma/blob/f2128424a4ae2276f597cb4069329e4d9b5dc1f1/.github/workflows/backport.yaml)
separately to start this off for us if this is what we settle on.
- Not totally sure if from this workflow we can tell renovate to do
something, maybe we just add something to the description of the PR made
here with a link to the dependency dashboard and some text saying "[ ]
Ask renovate to perform maintenance on the lock file" **edit**: we added
this


Let me know your thoughts on the above, other than that I think the only
way to test this properly is to temporarily manually change our
active-branches.json, merge to main, then run this to see if it updates
as intended. I'd also be curious to know what happens if there are no
changes here, I would guess it would just no make the PR.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>