chore(x): adds XDl component (#4228)
Similar to https://github.com/kumahq/kuma-gui/pull/4191 this PR adds a
thin layer over native `dl` so we can provide our own styles to a `dl`
in a more controlled way that adding global styles and or having a class
that we add to every dl.

Like XTable, this component should be seen in atomic terms as an atom,
i.e. the smallest possible purpose (its almost a native `dl`). In
avoiding global styles and using a restricted set of `:variants` we can
have more control over `dl`s globally across our application.

As with XTable this PR includes a single implementation, once this is
merged we can roll it out everywhere else.

My only slight concern about this PR is the awkward three letter name:
`XDl`, however this stick consistently to the pattern of:

1. All `x` components begin with `X` so we never hit vue restrictions
over disallowing single word components.
2. These "atom" components (which are a super thin layer of native
elements) use the exact same name as the native element so its
recognisable to other engineers - this makes its super obvious to use
and know what it is/should be used for.

Given the above, there is no other name we can use but `XDl`

Signed-off-by: John Cowen <john.cowen@konghq.com>