chore: slight tweaks to post-upgrade/build targets for docs (#4274)
I made `post-upgrade` a generic target, so then we use this target and
we can add more than just an `msw` specific script.

I also noticed that help docs were broken after `build` because the help
text for `build` was added on the `.PHONY`, so I fixed that up also.

Signed-off-by: John Cowen <john.cowen@konghq.com>