chore(kuma-http-api): workflow for auto-sync (#4275)
This is the first part of having automatic updates on our OAS based
types from kuma. Until there is an agreement on wether this new workflow
is being triggered via `cron` or `repository_dispatch` event and for
further testing, I've left it as manual trigger.
The approach of leveraging `repository_dispatch` events requires a
workflow in kuma that runs on every push to the base branch. In case
there is a diff between HEAD^ and HEAD in `docs/generated/openapi.yaml`
it should perform a repository dispatch to the kuma-gui repo, which will
then be picked up by this new workflow.
If this is not suitable we want to run this workflow on a regular basis
via `cron` (probably daily), but this means that we may be off sync and
no sync PR is being created until the next iteration.

Part of https://github.com/kumahq/kuma-gui/issues/2876

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>