feat(data-planes): relocate xds connections to new tab (#4268)
Moves the `XDS connections` to a new tab. With having in mind that we
want to align it with a zone's `KDS connections`
(https://github.com/kumahq/kuma-gui/issues/4237), I've created a new
view `SubscriptionsListView` in `subscriptions`.

In order to make it reusable for both `data-planes` and `zones` I've
added two props on the route level: `i18nPrefix` and `routePrefix`. In
the future we'd like to pass this down to any child routes from the root
route of each module.

Closes https://github.com/kumahq/kuma-gui/issues/4264

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>