chore: integrate XDl into DataPlaneDetailView (#4276)
https://github.com/kumahq/kuma-gui/pull/4228 added a new XDl component,
one of the reasons for having this was to have a simpler/more-grokkable
way to author `dl` type lists.

One area of code where I thought it might prove helpful was our new
DataPlaneDetailView.vue, so I figured I'd do an initial rollout here to
get a feel for it. Once I'd done this I kinda went a bit further after
realizing a few things, and because of this I've split this PR up into
commits so its easier to see the other refactors/reorgs.

All in all we've discussed this page a little in regards to it being a
little more grokkable and I wanted to see if it could be improved
without resorting to splitting things up into a number of smaller
single-use somponents.

Rough list of changes:

- Move DefinitionCard usage to XDl
61a6f6412886027bcaa185facb53440b74277faa
- Moved some data loading our of this view and into the drawer where its
used ee092b489fc299a20768c0782656ed19e568d93c
- Reorg/refactor the data interactions in this page using our composable
DataSource/DataLoader to move the data and loaders closer to where its
needed bf330813198b75257988f8638e58b3cdedf8b2d7
77cef1f95dd397055010693fab3ca9879a83a383

Its easier to do these sorts of reorgs when you aren't trying to build a
new feature i.e. hindsight is everything, plus I might have totally
missed something that I'm not aware of here. If necessary we can pull it
back to just the DefinitionCard > XDl migration. Although I really do
think all of this combined is a big improvement, and in doing this there
are some other things I spotted we could do.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>