fix(@kumahq/settings): fix can/env typing (#4288)
Unless I'm mistaken, I noticed that our typing for can/env wasn't
working as well as I remembered the first attempt to be. Therefore I've
reverted it back to the first attempt and added some `ts-expect-errors`
for regression assertions in the tests.

These are still open for improvement if there're are suggestions/ideas
for an alternative, but the `ts-expect-errors` will help us do that
without losing the typing we want here.

I was a bit surprised we regressed here, as I was totally sure we'd
figured this out, so if I'm mistaken in anyway please let me know!

Signed-off-by: John Cowen <john.cowen@konghq.com>