refactor(zones,-ingresses,-egresses): relocate subscriptions listing (#4289)
Aligns everything around subscriptions with dataplane subscriptions by
relocating the listing to a new tab (KDS connections/XDS connections)
and adding an info about the current active connection to the respective
about section.
Affects all `zones`, `zone-ingresses` and `zone-egresses`.
In order to reuse the `SubscriptionsListView` I had to do some data
modeling to align the data of the respective subscriptions of
`zoneInsight`, `dataplaneInsight`, `zoneIngressInsight` and
`zoneEgressInsight`:
- `instance.id`: either `subscriptions[].controlPlaneInstanceId` for DPs
or `subscriptions[].zoneInstanceId` for zones
- `instance.version`: either `subscriptions[].version.kumaDp.version`
for DPs or `subscriptions[].version.kumaCp.version` for zones

---

While I was revisiting `zones`, `zone-ingresses` and `zone-egresses`
I've also taken the chance to replace all `DescriptionCard`s with the
new `XDl` component.

Closes https://github.com/kumahq/kuma-gui/issues/4237

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>