feat: create XErrorState (#4299)
Adds `XErrorState` in favor of `ErrorBlock` and removes `ErrorBlock`.
Additionally the `ApiError` class now implements `AipError` which comes
from the OAS and is based on [Kong AIP
Error](https://kong-aip.netlify.app/aip/193/).

Further thoughts:
- `type` and `detail` fields are currently not required by the OAS of
errors, but it should be. I think `detail` is one of the most important
fields because it should provide human readable details about the error.
~I'll create an issue for this.~ There is an issue about this already
https://github.com/kumahq/kuma/issues/12415
- error `title` and `detail` are currently not translated and rendered
as is. For now this is not a big issue because we only have one language
(`en`) and both fields are set with english texts. However in the future
we might provide more than one language. In this case we might need to
support some additional cases or fields including keys that serve as
paths to a translation or the API responds with in correct language.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>