chore(x): uses memo/nano-container to ensure x injections as singletons (#4328)
I noticed that our shiki based syntax highlighter has always been
called/instantiated for every single instance of the component that we
use.

You can test this out by dropping a console log inside of
syntaxHighligher and navigating to various codeblock pages previous to
this PR.

This PR uses a very smol memo-like nano-container for lazy singletons
for the creation of `x`s `use*` getters thus making sure that the
highlighter function is only ever called once.

Once this PR is applied you'll see a console log only prints once.

Notes:

- Importantly I want this pattern to be as small and copy/pastable as
possible, its likely to be reused but I don't want to add a level of
dependencies to all the Vue plugins we are likely to make in the future.
- I also want this pattern to be interchangeable with a larger
container, hence the reliance on using native objects (map and symbol)
and "typed uris" is important here

Signed-off-by: John Cowen <john.cowen@konghq.com>