feat(data-planes/traffic/summary): use kri for policy origin link (#4315)
With the decision made in
https://github.com/kumahq/kuma/blob/master/docs/madr/decisions/086-endpoint-to-retrieve-a-resource-by-kri.md
we will soon be able to retrieve a resource by its KRI. Currently the
link from a dataplane traffic summary policy origin to a policy is
broken as we are not able to reconstruct the exact parameters from the
given `kri` to retrieve the resource. With the expected changes we are
able to just use the `kri` as is and fetch the policy data for the
`PolicyDetailView`.

---

I've tried several options to get the mocks for the generic kri endpoint
running:

1. Have one mock handler in `fs` that maps other mock handlers from the
`_resources`-endpoint and `shortName`s that it returns to respective
handlers dynamically. This wasn't possible as for now we cannot turn the
mock handlers to async due to limitations in cypress interceptors.
2. Use individual
`/_kri/kri/<shortName>/:mesh/:zone/:namespace/:name/:sectionName`
semi-generic handlers for different resource types in `fs` that directly
map to respective handlers. The URL transformation in `fake-api` worked
well (`/_kri/kri_<shortName>_:mesh_:zone_:namespace_:name_:sectionName`
to `/_kri/kri/<shortName>/:mesh/:zone/:namespace/:name/:sectionName`),
but unfortunately at least msw couldn't handle optional parameters. A
KRI segment may be empty (`''`) which would lead to URLs like
`/_kri/kri/<shortName>/////`.
3. Map individual semi-generic KRI handlers to respective handlers in
`fs` and preserve the format of the KRIs like
`/_kri/kri_<shortName>_:mesh_:zone_:namespace_:name_:sectionName`.
Unfortunately this failed with msw as it expects the URL separator to be
`/` (not `_`) and doesn't allow multiple path parameters per segment.
4. The chosen approach for now is to map individual semi-generic KRI
handlers like `/_kri/kri_<shortName>_:kri` to respective handlers and
consider a `kri` parameter in each of these handlers by parsing it if
available, otherwise fall back to already used path parameters (i.e.
`request.params.name`).

Closes https://github.com/kumahq/kuma-gui/issues/4310

Related backend PR: https://github.com/kumahq/kuma/pull/14675

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>